Add files via upload
Anomaly detection using autoencoding model. This is one of the first topics I worked on. 
Here with have collected data from sensors. Some observations present anomalies, which we are trying to figure out using our autoencoding model. 
In order to do that, we learned from a training set with normal observations during a chosen period . Next we tested it with a short period of anomalies and checked the results.